[BEAM-9432] Move expansion service into its own project. (#11035)
