+++ sdks/java/expansion-service/build.gradle
@@ -17,11 +17,10 @@
  */
 
 plugins { id 'org.apache.beam.module' }
-applyJavaNature(automaticModuleName: 'org.apache.beam.runners.core.construction')
+applyJavaNature(automaticModuleName: 'org.apache.beam.sdk.expansion.service')
 
-description = "Apache Beam :: Runners :: Core Construction Java"
-ext.summary = """Beam Runners Core provides utilities to aid runner authors interact with a Pipeline
-prior to execution."""
+description = "Apache Beam :: SDKs :: Java :: Expansion Service"
+ext.summary = """Contains code that can be used to run an expansion service."""
 
 
 // Exclude tests that need a runner
@@ -34,22 +33,18 @@ test {
 
 dependencies {
   compile project(path: ":model:pipeline", configuration: "shadow")
-  compile project(path: ":model:job-management", configuration: "shadow")
+  compile project(path: ":model:fn-execution", configuration: "shadow")
   compile project(path: ":sdks:java:core", configuration: "shadow")
+  compile project(path: ":runners:core-construction-java")
   compile library.java.vendored_grpc_1_26_0
   compile library.java.vendored_guava_26_0_jre
-  compile library.java.classgraph
-  compile library.java.jackson_core
-  compile library.java.jackson_databind
-  compile library.java.joda_time
   compile library.java.slf4j_api
-  testCompile library.java.hamcrest_core
-  testCompile library.java.hamcrest_library
+  runtimeOnly library.java.slf4j_jdk14
   testCompile library.java.junit
-  testCompile library.java.mockito_core
-  testCompile library.java.jackson_annotations
-  testCompile library.java.jackson_dataformat_yaml
-  testCompile project(path: ":model:fn-execution", configuration: "shadow")
-  testCompile project(path: ":sdks:java:core", configuration: "testRuntime")
-  testRuntimeOnly library.java.slf4j_jdk14
+}
+
+task runExpansionService (type: JavaExec) {
+  main = "org.apache.beam.sdk.expansion.service.ExpansionService"
+  classpath = sourceSets.main.runtimeClasspath
+  args = [project.findProperty("constructionService.port") ?: "8097"]
 }
