+++ sdks/java/expansion-service/src/main/java/org/apache/beam/sdk/expansion/service/ExpansionServer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.core.construction.expansion;
+package org.apache.beam.sdk.expansion.service;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
