+++ sdks/java/expansion-service/src/main/java/org/apache/beam/sdk/expansion/service/package-info.java
@@ -17,4 +17,4 @@
  */
 
 /** Classes used to expand cross-language transforms. */
-package org.apache.beam.runners.core.construction.expansion;
+package org.apache.beam.sdk.expansion.service;
