+++ sdks/java/testing/expansion-service/build.gradle
@@ -28,11 +28,12 @@ dependencies {
   compile project(path: ":runners:core-construction-java")
   compile project(path: ":sdks:java:io:parquet")
   compile project(path: ":sdks:java:core", configuration: "shadow")
+  compile project(":sdks:java:expansion-service")
   testRuntime library.java.hadoop_client
 }
 
 task runTestExpansionService (type: JavaExec) {
-  main = "org.apache.beam.sdk.testing.expansion.TestExpansionService"
+  main = "org.apache.beam.sdk.expansion.service.ExpansionService"
   classpath = sourceSets.test.runtimeClasspath
   args = [project.findProperty("constructionService.port") ?: "8097"]
 }
@@ -44,9 +45,10 @@ task buildTestExpansionServiceJar(type: ShadowJar) {
   mergeServiceFiles()
   manifest {
     attributes(
-            'Main-Class': 'org.apache.beam.sdk.testing.expansion.TestExpansionService'
+            'Main-Class': 'org.apache.beam.sdk.expansion.service.ExpansionService'
     )
   }
+  exclude 'META-INF/*.RSA', 'META-INF/*.SF','META-INF/*.DSA'
   from { project.configurations.testRuntime.collect { it.isDirectory() ? it : zipTree(it) }}
   from sourceSets.main.output
   from sourceSets.test.output
