+++ settings.gradle
@@ -65,6 +65,7 @@ include ":sdks:java:bom"
 include ":sdks:java:build-tools"
 include ":sdks:java:container"
 include ":sdks:java:core"
+include ":sdks:java:expansion-service"
 include ":sdks:java:extensions:euphoria"
 include ":sdks:java:extensions:kryo"
 include ":sdks:java:extensions:google-cloud-platform-core"
